requirements/development.txt
-r base.txt
django-debug-toolbar>=1.8
Werkzeug>=0.12.2
ipdb==0.10.3
django-extensions==1.9.0
django-test-without-migrations==0.6
ipython[notebook]
